+++ packages/cli/templates/Gjs/misc/fileUtils.js
@@ -1,0 +1,6 @@
+<% if(moduleType === 'esm') { %>
+    export imports.misc.fileUtils;
+<% } else { %>  
+    module.exports = imports.misc.fileUtils;
+<% } %>
+