+++ packages/cli/templates/Gjs/misc/gnomeSession.js
@@ -1,0 +1,6 @@
+<% if(moduleType === 'esm') { %>
+    export imports.misc.gnomeSession;
+<% } else { %>  
+    module.exports = imports.misc.gnomeSession;
+<% } %>
+